Added new events for TRailblazers.

+++ EliteJournalReader/Events/PowerplayEvent.cs
@@ -16,7 +16,7 @@ namespace EliteJournalReader.Events
         {
             public string Power { get; set; }
             public int Rank { get; set; }
-            public int Merits { get; set; }
+            public ulong Merits { get; set; }
             public int Votes { get; set; }
             public long TimePledged { get; set; }
         }
